Add `ChainWorkerState::process_confirmed_block`
Refactor to move the code from the `WorkerState` into the
`ChainWorkerState`.
